+++ src/main/java/net/quickwrite/fluent4j/ast/placeable/base/FluentSelectable.java
@@ -1,0 +1,4 @@
+package net.quickwrite.fluent4j.ast.placeable.base;
+
+public interface FluentSelectable {
+}
